requirement.txt
langchain
langchain_google_genai
langchain_core
pandas
fastapi
uvicorn
python-dotenv